+++ CHANGELOG.md
@@ -487,6 +487,7 @@
 - Work out a Budget's period based on the financial year
 - Users can see Current activities and Historic activities in different tabs
 - Reorder reports in a more intuitive manner
+- Group activities by hierarchy on the view of a single report page
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-30...HEAD
 [release-30]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-29...release-30